Optimizations
Worked on optimizing some things about the first person camera and the movement of the player because there was a bug that was annoying. Also decided to change some thigns about the main game like story and what to work on. Started adding animations and a character model. Will start on building the levels and then decide what objects can be controlable and not.