Level Building
Started doing animations and scrpts for intereacting with objects. Working on the first level and got a really good and new idea on how I want to approach the mechanic of the game and how it will work. Will start building on that while making the leves. Also came with a cool idea for story and adding cutscenes.