Final Changes + Testing
As mentioned in the last commit, I returned to add the last parts of the project.  All of I had to add was the ending of the final lelvel and make a main menu where you can switch between levels. I also tested to see if everything works as intended and modified where I found bugs. The last thing I did was removing some comments and debug logs from the game.